Remove invite in rule 8
No other rule has an invite link in it, there's no reason to include an invite link to a server that discord.js has no affiliation with and no other relation to.  The embed adds clutter to the rules and is generally inconsistent with the rest of the rules' style.